Added %last_damage_taken% placeholders for receive hit activators
It's different from the %last_damage_taken% from SCore because the one in SCore records the value after the event has passed.
These placeholders that were introduced in this commit provide the supposed damage so even if the event is cancelled, users can extract the damage they're supposed to take to adjust their logic accordingly.